[chore] `LottoStatics` 클래스의 `divideOfListCount` 메서드 파라미터 long 형으로 들어오도록 변경

+++ src/main/java/lotto/domain/LottoStatics.java
@@ -20,7 +20,7 @@ public class LottoStatics {
                 .reduce(BigDecimal.ZERO, BigDecimal::add);
     }
 
-    private static BigDecimal divideOfListCount(BigDecimal sum, int count) {
+    private static BigDecimal divideOfListCount(BigDecimal sum, long count) {
         return sum.divide(BigDecimal.valueOf(UNIT_OF_ACCOUNT * PERCENT * count), 2, RoundingMode.HALF_EVEN);
     }
